+++ applanga.py
@@ -17,6 +17,7 @@ cli.add_command(commands.config.config)
 cli.add_command(commands.init.init)
 cli.add_command(commands.pull.pull)
 cli.add_command(commands.push.push)
+cli.add_command(commands.updateSettingsfiles.updateSettingsfiles)
 cli.add_command(commands.pullSource.pullSource)
 cli.add_command(commands.pushTarget.pushTarget)