+++ commands/__init__.py
@@ -2,5 +2,6 @@ from commands import config
 from commands import init
 from commands import pull
 from commands import push
+from commands import updateSettingsfiles
 from commands import pullSource
 from commands import pushTarget